# Conflicts:
#	Directory.Packages.props
#	test/Nerdbank.MessagePack.TUnit/ArraysOfPrimitivesTests.cs
#	test/Nerdbank.MessagePack.TUnit/ConvertToJsonTests.cs
#	test/Nerdbank.MessagePack.TUnit/DerivedTypeTests.cs
#	test/Nerdbank.MessagePack.TUnit/MessagePackReaderTests.cs
#	test/Nerdbank.MessagePack.TUnit/MessagePackSerializerTests.cs
#	test/Nerdbank.MessagePack.TUnit/NativeAOTTests.cs
#	test/Nerdbank.MessagePack.TUnit/ObjectsAsArraysTests.cs
#	test/Nerdbank.MessagePack.TUnit/OptionalConvertersTests.cs
#	test/Nerdbank.MessagePack.Tests/BuiltInConverterTests.cs
#	test/Nerdbank.MessagePack.Tests/StructuralEqualityComparerTests.cs

+++ src/Nerdbank.MessagePack/MessagePackAsyncReader.cs
@@ -26,7 +26,7 @@ public class MessagePackAsyncReader(PipeReader pipeReader) : IDisposable
 	private bool readerReturned = true;
 
 	/// <inheritdoc cref="MessagePackStreamingReader.ExpectedRemainingStructures"/>
-	private uint expectedRemainingStructures;
+	private ulong expectedRemainingStructures;
 
 	/// <summary>
 	/// Gets a cancellation token to consider for calls into this object.
